Complete rewrite of DragToPop
